requirements.txt
pydantic
anthropic
python-dotenv
